SOC-3193 | refactor, extracting model function to mixins
